Add json/jsonl io functions
